DO NOT MERGE
but it works, kinda.

Loads process start/stop from the process monitor driver.
Sends via GRPC over to process agent.

Probably doesn't compile on linux.

Would need some thought on how to combine the logic to share as much as
possible, vs. the copy/paste that's done here.
