+++ cmd/process-agent/windows_resources/process-agent.rc
@@ -7,7 +7,7 @@
 
 #define MAIN_ICON 1
 
-MAIN_ICON ICON "..\\..\\..\omnibus\\resources\\agent\\msi\\assets\\project.ico"
+MAIN_ICON ICON "..\\..\\..\\omnibus\\resources\\agent\\msi\\assets\\project.ico"
 
 #define VS_VERSION_INFO 1
 VS_VERSION_INFO VERSIONINFO
