+++ cmd/system-probe/modules/all_windows.go
@@ -20,6 +20,7 @@ import (
 // All System Probe modules should register their factories here
 var All = []module.Factory{
 	NetworkTracer,
+	SecurityRuntime,
 }
 
 const (
